Make CDP client key optional
