+++ typescript/packages/x402-express/src/index.ts
@@ -26,6 +26,7 @@ import { useFacilitator } from "x402/verify";
  * @param payTo - The address to receive payments
  * @param routes - Configuration for protected routes and their payment requirements
  * @param facilitator - Optional configuration for the payment facilitator service
+ * @param cdpClientKey - Optional CDP client API key (required for built-in paywall)
  * @returns An Express middleware handler
  *
  * @example
@@ -65,6 +66,7 @@ export function paymentMiddleware(
   payTo: Address,
   routes: RoutesConfig,
   facilitator?: FacilitatorConfig,
+  cdpClientKey?: string,
 ) {
   const { verify, settle } = useFacilitator(facilitator);
   const x402Version = 1;
@@ -140,6 +142,7 @@ export function paymentMiddleware(
             >[0]["paymentRequirements"],
             currentUrl: req.originalUrl,
             testnet: network === "base-sepolia",
+            cdpClientKey,
           });
         res.status(402).send(html);
         return;
