+++ typescript/packages/x402-hono/src/index.ts
@@ -26,6 +26,7 @@ import { useFacilitator } from "x402/verify";
  * @param payTo - The address to receive payments
  * @param routes - Configuration for protected routes and their payment requirements
  * @param facilitator - Optional configuration for the payment facilitator service
+ * @param cdpClientKey - Optional CDP client API key (required for built-in paywall)
  * @returns A Hono middleware handler
  *
  * @example
@@ -65,6 +66,7 @@ export function paymentMiddleware(
   payTo: Address,
   routes: RoutesConfig,
   facilitator?: FacilitatorConfig,
+  cdpClientKey?: string,
 ) {
   const { verify, settle } = useFacilitator(facilitator);
   const x402Version = 1;
@@ -135,6 +137,7 @@ export function paymentMiddleware(
             >[0]["paymentRequirements"],
             currentUrl,
             testnet: network === "base-sepolia",
+            cdpClientKey,
           });
         return c.html(html, 402);
       }
