+++ typescript/packages/x402/src/paywall/src/window.d.ts
@@ -1,0 +1,22 @@
+import { PaymentRequirements } from "../../types/verify";
+
+declare global {
+  interface Window {
+    x402: {
+      amount?: number;
+      testnet?: boolean;
+      paymentRequirements: PaymentRequirements | PaymentRequirements[];
+      currentUrl: string;
+      cdpClientKey?: string;
+      config: {
+        chainConfig: Record<
+          string,
+          {
+            usdcAddress: string;
+            usdcName: string;
+          }
+        >;
+      };
+    };
+  }
+}
