+++ typescript/site/middleware.ts
@@ -4,6 +4,7 @@ import { paymentMiddleware, Resource, Network } from "x402-next";
 const address = process.env.RESOURCE_WALLET_ADDRESS as Address;
 const network = process.env.NETWORK as Network;
 const facilitatorUrl = process.env.NEXT_PUBLIC_FACILITATOR_URL as Resource;
+const cdpClientKey = process.env.NEXT_PUBLIC_ONCHAINKIT_API_KEY;
 
 export const middleware = paymentMiddleware(
   address,
@@ -19,6 +20,7 @@ export const middleware = paymentMiddleware(
   {
     url: facilitatorUrl,
   },
+  cdpClientKey,
 );
 
 // Configure which paths the middleware should run on
